Change test class names and move helper classes
