+++ test/Microsoft.Azure.Devices.Edge.Test/Device.cs
@@ -8,12 +8,13 @@ namespace Microsoft.Azure.Devices.Edge.Test
     using Microsoft.Azure.Devices.Edge.Test.Common;
     using Microsoft.Azure.Devices.Edge.Test.Common.Certs;
     using Microsoft.Azure.Devices.Edge.Test.Common.Config;
+    using Microsoft.Azure.Devices.Edge.Test.Helpers;
     using Microsoft.Azure.Devices.Edge.Util;
     using NUnit.Framework;
     using NUnit.Framework.Interfaces;
     using Serilog;
 
-    class EndToEndWithCustomCertificates
+    class Device
     {
         CancellationTokenSource cts;
         IEdgeDaemon daemon;
