+++ test/Microsoft.Azure.Devices.Edge.Test/SetupFixture.cs
@@ -6,6 +6,7 @@ namespace Microsoft.Azure.Devices.Edge.Test
     using System.Threading;
     using System.Threading.Tasks;
     using Microsoft.Azure.Devices.Edge.Test.Common;
+    using Microsoft.Azure.Devices.Edge.Test.Helpers;
     using NUnit.Framework;
     using Serilog;
     using Serilog.Events;
