+++ test/Microsoft.Azure.Devices.Edge.Test/helpers/TestAdapterExtensions.cs
@@ -1,5 +1,6 @@
 // Copyright (c) Microsoft. All rights reserved.
-namespace Microsoft.Azure.Devices.Edge.Test
+
+namespace Microsoft.Azure.Devices.Edge.Test.Helpers
 {
     using System.Collections.Generic;
     using System.Linq;
