+++ test/Microsoft.Azure.Devices.Edge.Test/helpers/TestAuthenticationType.cs
@@ -1,7 +1,8 @@
 // Copyright (c) Microsoft. All rights reserved.
-namespace Microsoft.Azure.Devices.Edge.Test
+namespace Microsoft.Azure.Devices.Edge.Test.Helpers
 {
     using System.ComponentModel;
+    using Microsoft.Azure.Devices;
 
     public enum TestAuthenticationType
     {
